# Port CPU Trace from TH to ATen